+++ package/netlogin-nxu/files/usr/bin/netlogin_openwrt.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# This file is replaced during the package build from the repository root.
+# Keeping a source copy here makes the package directory self-describing.